Everything seems to work fine with the new way of setting the time steps.

+++ VisualPIC/DataReading/folderDataReader.py
@@ -162,7 +162,7 @@ class FolderDataReader:
             timeStep = int(file[-9:-3])
             timeSteps[i] = timeStep
             i+=1
-        return timeSteps
+        return timeSteps.astype(np.int64)
 
     def LoadHiPaceData(self):
         """HiPACE loader"""
